+++ CHANGELOG.md
@@ -14,6 +14,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Methods:
   - `OxidEsales\EshopCommunity\Core\UtilsView:parseThroughSmarty()`
   - `OxidEsales\EshopCommunity\Core\UtilsView:_smartyDefaultTemplateHandler()`
+  - `oxfunctions:ox_get_template()`
+  - `oxfunctions:ox_get_timestamp()`
+  - `oxfunctions:ox_get_secure()`
+  - `oxfunctions:ox_get_trusted()`
 
 ## [7.0.0-rc.2] - 2022-08-15